requirements.txt
yfinance 
mysql-connector-python
pandas 
matplotlib